requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pytest==8.1.1
ruff==0.3.5
pydantic-settings==2.2.1
starlette-context==0.3.6
Jinja2==3.1.3
SQLAlchemy==2.0.29
sqlalchemy-json==0.7.0
python-jose==3.3.0
passlib==1.7.4
sqlmodel==0.0.16
alembic==1.13.1
tenacity==8.2.3
autoflake
black==24.3.0
isort==5.13.2
mypy==1.9.0
pre-commit
coverage==7.4.4
types-python-jose
psycopg2-binary==2.9.9